hacked a cipher
